requirements.txt
gitpython
pygithub